+++ codex/serializers/opds/v2/progression.py
@@ -25,7 +25,7 @@ class OPDS2ProgressionLocationsSerializer(Serializer):
     """Progression Location Element."""
 
     fragments = ListField(child=CharField(read_only=True), read_only=True)
-    position = IntegerField(read_only=True)
+    position = IntegerField()
     progression = FloatField(read_only=True)
     total_progression = FloatField(read_only=True)