+++ codex/settings/__init__.py
@@ -71,6 +71,7 @@ RESET_ADMIN = not_falsy_env("CODEX_RESET_ADMIN")
 LOG_DIR = Path(environ.get("CODEX_LOG_DIR", CONFIG_PATH / "logs"))
 LOG_TO_CONSOLE = environ.get("CODEX_LOG_TO_CONSOLE") != "0"
 LOG_TO_FILE = environ.get("CODEX_LOG_TO_FILE") != "0"
+AUTH_REMOTE_USER = not_falsy_env("CODEX_AUTH_REMOTE_USER")
 THROTTLE_ANON = int(environ.get("CODEX_THROTTLE_ANON", "0"))
 THROTTLE_USER = int(environ.get("CODEX_THROTTLE_USER", "0"))
 THROTTLE_OPDS = int(environ.get("CODEX_THROTTLE_OPDS", "0"))
@@ -150,6 +151,7 @@ def _get_installed_apps():
         "whitenoise.runserver_nostatic",
         "django.contrib.staticfiles",
         "rest_framework",
+        "rest_framework.authtoken",
         "rest_registration",
         "corsheaders",
     ]
@@ -177,6 +179,10 @@ def _get_middleware():
         "django.middleware.common.CommonMiddleware",
         "django.middleware.csrf.CsrfViewMiddleware",
         "django.contrib.auth.middleware.AuthenticationMiddleware",
+    ]
+    if AUTH_REMOTE_USER:
+        middleware += ["codex.authentication.HttpRemoteUserMiddleware"]
+    middleware += [
         "django.contrib.messages.middleware.MessageMiddleware",
         "django.middleware.clickjacking.XFrameOptionsMiddleware",
         "codex.middleware.TimezoneMiddleware",
@@ -256,6 +262,11 @@ SESSION_ENGINE = "django.contrib.sessions.backends.cached_db"
 #   https://code.djangoproject.com/ticket/32674
 DEFAULT_AUTO_FIELD = "django.db.models.AutoField"
 
+if AUTH_REMOTE_USER:
+    AUTHENTICATION_BACKENDS = [
+        "django.contrib.auth.backends.RemoteUserBackend",
+        "django.contrib.auth.backends.ModelBackend",
+    ]
 
 # Password validation
 # https://docs.djangoproject.com/en/3.0/ref/settings/#auth-password-validators
@@ -340,6 +351,8 @@ for scope, value in _THROTTLE_MAP.items():
 REST_FRAMEWORK = {
     "DEFAULT_AUTHENTICATION_CLASSES": (
         "rest_framework.authentication.SessionAuthentication",
+        "rest_framework.authentication.TokenAuthentication",
+        "codex.authentication.BearerTokenAuthentication",
     ),
     "DEFAULT_RENDERER_CLASSES": (
         "djangorestframework_camel_case.render.CamelCaseJSONRenderer",