+++ codex/views/opds/auth.py
@@ -1,11 +1,21 @@
 """OPDS Authentican mixin."""
 
-from rest_framework.authentication import BasicAuthentication, SessionAuthentication
+from rest_framework.authentication import (
+    BasicAuthentication,
+    SessionAuthentication,
+    TokenAuthentication,
+)
 
+from codex.authentication import BearerTokenAuthentication
 from codex.views.auth import AuthMixin
 
 
 class OPDSAuthMixin(AuthMixin):
     """Add Basic Auth."""
 
-    authentication_classes = (BasicAuthentication, SessionAuthentication)
+    authentication_classes = (
+        SessionAuthentication,
+        BasicAuthentication,
+        TokenAuthentication,
+        BearerTokenAuthentication,
+    )